kernel: run unmodified static Linux x86-64 binaries (M0+M1)
An ELF with EI_OSABI == ELFOSABI_LINUX (3) loads with pcb.personality =
.linux; its `syscall` instructions route to a new translation layer
(cpu/syscall/linux.zig) instead of the native dispatch. doSyscall widens to
u64 and branches on personality before the native path; native binaries
(OSABI 0, the Zig default) are byte-for-byte unchanged. fork inherits it.

M0: write, exit, exit_group (freestanding raw-syscall binary).

M1: an unmodified static musl-libc program runs its full startup + printf.
  - syscalls: writev (musl's stdio flush path), ioctl -> -ENOTTY (the
    buffering/isatty probe), arch_prctl SET_FS/GET_FS (TLS bring-up,
    mirrors native sysSetTls), set_tid_address, getpid, rt_sigprocmask,
    set_robust_list.
  - elf_loader.setupLinuxInitialStack builds the SysV initial stack
    (argc/argv/envp/auxv, AT_PHDR..AT_RANDOM) in an eagerly-mapped top
    stack page and points the task's first-dispatch RSP at argc (16-aligned).
    The page is a normal anon frame; destroyAddressSpace reclaims it.

build.zig stages linux-bin/* (OSABI-stamped, non-PIE, linked >= 0x500000)
into /bin and rm -f's the ext2 image before genext2fs so adding a binary to
the tree forces a clean rebuild.

Boot-verified: a 876 KB `zig cc -target x86_64-linux-musl` binary prints
"hello from static musl libc on zigos! 2+2=4" through our libc-startup path;
0 unimpl syscalls, 0 panic. Reviewer: SAFE-TO-BOOT, 0 critical.

+++ src/cpu/syscall/entry.zig
@@ -167,7 +167,7 @@ fn wrmsr(msr: u32, value: u64) void {
 // The smp.zig assert guarantees the 8-byte read doesn't straddle a cache
 // line, so a same-CPU setTssRsp0 update is observed atomically here.
 
-extern fn doSyscall(num: u32, arg1: u32, arg2: u32, arg3: u32, frame: *anyopaque) callconv(.c) u32;
+extern fn doSyscall(num: u32, arg1: u32, arg2: u32, arg3: u32, frame: *anyopaque) callconv(.c) u64;
 
 fn CpuEntry(comptime cpu_id: u8) type {
     const off_user_rsp_str = std.fmt.comptimePrint("{d}", .{@as(usize, cpu_id) * 8});
